Migrated PoPSchema\Tags\WP\AbstractFunctionAPI
