+++ layers/Schema/packages/migrate-tags-wp/migrate/contractimplementations/cms-apis/load.php
@@ -1,5 +1,4 @@
 <?php
 
-require_once 'cms-functionapi.php';
 require_once 'cms-objectpropertyresolver.php';
 require_once 'cms-routingstate-hooks.php';
