+++ layers/Schema/packages/tags/src/TypeAPIs/TagTypeAPIInterface.php
@@ -15,4 +15,14 @@ interface TagTypeAPIInterface extends TaxonomyTypeAPIInterface
      * Indicates if the passed object is of type Tag
      */
     public function isInstanceOfTagType(object $object): bool;
+    public function getCustomPostTags($post_id, array $query = [], array $options = []): array;
+    public function getCustomPostTagCount($post_id, array $query = [], array $options = []): int;
+    public function getTag($tag_id);
+    public function getTagByName($tag_name);
+    public function getTags($query, array $options = []): array;
+    public function getTagCount(array $query = [], array $options = []): int;
+    public function getTagLink($tag_id);
+    public function getTagName($tag_id);
+    public function getTagBase();
+    public function setPostTags($post_id, array $tags, bool $append = false);
 }
